Refactor package names and imports in InfrastructureAgent code files
