+++ InfrastructureAgent/model/cpu_metrics.go
@@ -1,4 +1,4 @@
-package oneuptime_infrastructure_agent
+package model
 
 type CPUMetrics struct {
 	PercentUsed float64 `json:"percentUsed"`
