+++ InfrastructureAgent/model/disk_metrics.go
@@ -1,4 +1,4 @@
-package oneuptime_infrastructure_agent
+package model
 
 type BasicDiskMetrics struct {
 	Total       uint64  `json:"total"`
