+++ InfrastructureAgent/shutdown.go
@@ -1,4 +1,4 @@
-package oneuptime_infrastructure_agent
+package main
 
 import (
 	"os"
