+++ InfrastructureAgent/utils/agent.go
@@ -1,4 +1,4 @@
-package oneuptime_infrastructure_agent
+package utils
 
 import (
 	"encoding/json"
