+++ InfrastructureAgent/utils/disk.go
@@ -1,4 +1,4 @@
-package oneuptime_infrastructure_agent
+package utils
 
 import (
 	"log/slog"
